Move withTransaction to dedicated models/tx module
Refactors withTransaction function from web/federation.ts to models/tx.ts
and updates all imports to use the new module path. This improves code
organization by placing the database transaction utility in the models
layer where it belongs.
