+++ web/routes/@[username]/[idOrYear]/[slug]/quotes.tsx
@@ -11,12 +11,12 @@ import type {
   PostMedium,
   Reaction,
 } from "@hackerspub/models/schema";
+import { withTransaction } from "@hackerspub/models/tx";
 import * as v from "@valibot/valibot";
 import { sql } from "drizzle-orm";
 import { Msg } from "../../../../components/Msg.tsx";
 import { PostExcerpt } from "../../../../components/PostExcerpt.tsx";
 import { db } from "../../../../db.ts";
-import { withTransaction } from "../../../../federation.ts";
 import { ArticleExcerpt } from "../../../../islands/ArticleExcerpt.tsx";
 import { Composer } from "../../../../islands/Composer.tsx";
 import { PostControls } from "../../../../islands/PostControls.tsx";
