+++ web/routes/@[username]/[idOrYear]/quotes.tsx
@@ -15,13 +15,13 @@ import type {
   PostMedium,
   Reaction,
 } from "@hackerspub/models/schema";
+import { withTransaction } from "@hackerspub/models/tx";
 import { validateUuid } from "@hackerspub/models/uuid";
 import * as v from "@valibot/valibot";
 import { sql } from "drizzle-orm";
 import { Msg } from "../../../components/Msg.tsx";
 import { PostExcerpt } from "../../../components/PostExcerpt.tsx";
 import { db } from "../../../db.ts";
-import { withTransaction } from "../../../federation.ts";
 import { Composer } from "../../../islands/Composer.tsx";
 import { PostControls } from "../../../islands/PostControls.tsx";
 import { define } from "../../../utils.ts";
